+++ libephemeris/__init__.py
@@ -756,7 +756,7 @@ from .arabic_parts import calc_all_arabic_parts
 # .env file loader (public API for manual reloading)
 from ._dotenv import load_dotenv
 
-__version__ = "1.0.0a6"
+__version__ = "1.0.0a7"
 version = __version__
 __author__ = "Giacomo Battaglia"
 __license__ = "AGPL-3.0"